Hide implementation details from validation code
