+++ src/server/messages.ts
@@ -1,6 +1,5 @@
-import * as AdmZip from 'adm-zip';
 import { default as fetch }from 'node-fetch';
-import { createMultipart, createResponseMessage, extractLogs, getResponseCode } from './mime';
+import { createMultipart, createResponseMessage, extractLogs, getMultipartBody, getResponseCode } from './mime';
 import { ServerResult } from '../client/server_result';
 import * as vscode from 'vscode';
 import { log } from '../logger';
@@ -16,29 +15,11 @@ import { log } from '../logger';
 */
 
 export async function validate(filename: string, project: string, text: string): Promise<ServerResult> {
-    // First create the body of the message, since we'll need some information
-    // from it to create the headers
-    const zip = new AdmZip();
-    // text.length does not account for the encoding, so using that will allocate
-    // less memory that required and truncate the text in the zip!
-    zip.addFile(`00atf/${filename}`, Buffer.alloc(Buffer.byteLength(text), text));
-    // Use the ISO-8859-1 encoding, to ensure we don't misinterpret the zip contents.
-    const encodedText = zip.toBuffer().toString('latin1');
     // TODO replace this with appropriate commands and reponse ID params
-    const fullMessage = createMultipart("atf", filename, project, encodedText);
+    const fullMessage = createMultipart("atf", filename, project, text);
     log('info', fullMessage.toString());
-    // Build the message manually so we can control the encoding of the zip
-    // This is mimicking what the mimemessage package does
+    const body = getMultipartBody(fullMessage);
     const boundary = fullMessage.contentType().params.boundary;
-    const sep = "\r\n";
-    const body: Buffer = Buffer.concat([
-        Buffer.from('--' + boundary + sep),
-        // We may not need this? Holdover from Nammu (changing all line endings to \r\n)
-        Buffer.from(fullMessage._body[0].toString().replace(/\r\n/g, "\n").replace("\n", "\r\n")),
-        Buffer.from('\r\n' + '--' + boundary + sep),
-        Buffer.from(fullMessage._body[1].toString(), 'latin1'),
-        Buffer.from(sep + '--' + boundary + '--')
-    ]);
 
     const host = "http://build-oracc.museum.upenn.edu";
     const port = 8085;
@@ -51,7 +32,7 @@ export async function validate(filename: string, project: string, text: string):
         // TODO: Should this be the whole body or only part? Compare with Nammu.
         'Content-Length': String(Buffer.byteLength(body))
     }
-    
+
     const options = {
         method: 'POST',
         headers: headers,
@@ -88,7 +69,7 @@ export async function validate(filename: string, project: string, text: string):
         log('error', `A problem has occurred: ${err}`);
         return;  // Should return an empty result?
     }
-    
+
     return new ServerResult(finalResult.get('oracc.log'),
                             finalResult.get('request.log'));
 }
